+++ changelog.md
@@ -13,3 +13,33 @@ added - "walk" as a main menu option
 added - movement() function
 added - options to walk's sub menu "N,S,E,W" in movement()
 fixed - movement so player can not walk off "map"
+
+v2.0 GitHub Name: Map---part-2 
+created - empty list to function as an inventory of objects  
+created - objects without a dictionary 
+added - conditional statements to deal with objects
+
+v3.0 GitHub Name: Map---Dictionarys Name: RPG: Castal Adventure!  
+added - dictionary for tile/map attributes (description strings) 
+fixed - room description for main menu updated - object data to a dictionary  
+added - 'look' as a main menu option  
+added - inspectRoom() function to deal with objects in the game  
+added - a fuction to deal with the chest object  
+added - a function to deal with the key object  
+added - docStrings for each of the functions
+
+v4.0 GitHub Name: Map---Librarys Name: RPG: Castal Adventure!  
+added - library random so the key can generate in a random room  
+added - libary sys so that 'quit' options more efficently end the game removed - while loop variables  
+added - messages dictionary for repetative messages in the game 
+fixed - DocString line ups fixed - Headers DocString 
+updated - Header information  
+created - lists for the main menu and directions menu 
+updated - user input to deal ignor capitalization
+
+v5.0  
+added - module for tile data 
+remove - unessesary map and tile global variables calls  
+added - call values using the module_name module_name.function_name() fixed - error and quit messages to be called more efficently  
+added - Treasure object.  
+added - Treasure function to end the game.